+++ app/youtube/page.tsx
@@ -4,44 +4,7 @@ import { useState } from "react";
 import { BackHeader } from "../components/BackHeader";
 import { PageFooter } from "../components/PageFooter";
 import { ClearButton } from "../components/ClearButton";
-
-function CopyButton({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false);
-
-  const copyToClipboard = async () => {
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    } catch (err) {
-      console.error("Failed to copy text:", err);
-    }
-  };
-
-  return (
-    <button
-      onClick={copyToClipboard}
-      className="px-3 py-1 text-sm rounded-md bg-blue-500 hover:bg-blue-600 
-               text-white transition-colors duration-200 flex items-center gap-2"
-    >
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        className="h-4 w-4"
-        fill="none"
-        viewBox="0 0 24 24"
-        stroke="currentColor"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={2}
-          d="M8 16H6a2 2 0 01-2-2V6a2 2 0 012-2h8a2 2 0 012 2v2m-6 12h8a2 2 0 002-2v-8a2 2 0 00-2-2h-8a2 2 0 00-2 2v8a2 2 0 002 2z"
-        />
-      </svg>
-      {copied ? "Copied!" : "Copy"}
-    </button>
-  );
-}
+import { CopyButton } from "../components/CopyButton";
 
 function Transcript({ text }: { text?: string }) {
   if (!text) {
